pyproject.toml
[project]
name = "aind-experiment-vr-foraging-open-ephys"
version = "0.1.0"
description = "A thin wrapper that runs VrForaging + OpenEphys experiments"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aind-behavior-vr-foraging[launcher]",
    "aind-clabe @ git+https://github.com/AllenNeuralDynamics/clabe@improve-app-api",
]

[dependency-groups]

dev = [
    "ruff",
    "codespell"
]

[tool.uv.sources]
aind-behavior-vr-foraging = { path = "./Aind.Behavior.VrForaging" }


[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock,*Generated.cs'
ignore-words-list = 'nd'

[tool.uv]
default-groups = ['dev']
required-version = '>=0.8.4'